Reduce duplication of register lists
There are several (partial) copies of the holding register lists.
Define that the copy in inverter.py is the master copy. Extend
it so that writable registers specify the range of valid values
which may be written.

There is an asymmetry in that when reading attributes,
they are scaled and converted to a friendly form, potentially
combining multiple holding registers, whereas writes are currently
expected to write a raw uint16 to a single register. For now,
this requires that single-register aliases are added for
multi-slot attributes ((primarily the Timeslot ones).

lut.py seemed unused, that can just be removed.

pdu/write_registers.py had a list of writable registers, but
I think it should just encode what it's been told to, rather than
making policy decisions. So just remove that. If it is needed,
a list can be generated at startup from the master copy.

commands.py had its own list of writable registers. Plus the
individual commands were doing explicit value range-range checks.
Introduce a pair of helper functions to replace this scheme:
 Commands.write_named_register()
 Inverter.lookup_writable_register()

Fix a typo in register.py in the fn used for converting
the system time.