sync with givenergy-local step 1 - whitespace
givenergy-local merged a copy of the code into their
tree at this point. Lots of changes made before the initial
commit.

Step 1 is just running most of the code through the 'black' util
with standard line length and quote processing. So it should
be a no-op change.

Some files seem to have been excluded from the string processing:
  lut.py (appears to be no longer used)
  code.py
  framer.py
  __init__.py

Was it just top-level py files that escaped ?

Manually wrap a few import lines that aren't long in this
tree, but become long when 'custom_component.givenergy-local.' was
prepended to all the import locations.