Remove dependency on pydantic.
When pydantic instantiates an Inverter or Battery instance,
it fully populates all the attributes, converting registers
as required. This is wasteful for atributes that are never
accessed, and scales badly as the number of registers increases,
and as aliases are added to access registers in different formats.

Reimplement Inverter and Battery to be straightforward classes
which lazily convert register accesses. Accesses to
inverter.x turn into a call to inverter.get('x') which actually
performs the conversion at that time.

Also, change the default value for the register cache to None
rather than 0, so that it's possible to distinguish a missing
value from a genuinely 0 value.

The interface is almost unchanged. The only noticable difference
is that if you hold onto an Inverter instance across a
plant refresh: previously, the attributes would be unchanged,
whereas now, any subsequent access will reflect the updated
values.  And it's now more efficient to copy an inverter attribute
into a local variable than to repeatedly access it.