Add some holding register aliases for start/end of timeslots.
In order to be able to write timeslots, we need to be able to
address the start and end registers directly. Use a script to
add a start and end register alias for each timeslot pair,
with valid settings to mark them writable.

Also add valid=(4, 100)  to charging soc registers.