add client option to record raw socket data to a file
For debugging, it's useful to be able to record the bytestream
to a file.  An an optional recorder param to Client(), and
it will record all bytes to that. Can be a simple file or
a GzipFile, for example. (There's a lot of repetition, so
it compresses quite well.)

Remove some old commented-out code which looked like it
was storing the data in a queue, with a task to flush it
out from time to time.