Flesh out the ReadBatteryInputRegisters implementation
Recently I've been seeing NotImplemented exceptions due to unknown
frame type #22, which seems to be "ReadBatteryInputRegisters".
There was some partial support in place, but it appeared to
be incomplete. Try to complete it in a consistent way.

I don't have a definition of the registers, so this does
little more than stop exceptions being thrown.