Reimplement (and simplify) the pdu module.
There were lots of intermediate classes making up the
various transparent Request/Reply messages, but seemed like a lot of
extra fuss. In particular, the rules for how to encode and
decode the various bits of messages was distributed throughout the code,
making it all quite hard to follow.

Recognise that they're all doing largely the same thing, with just details about
whether they include a count and/or a values field, so switch to a data-driven
rather than code-driven model: all the work is done in the Transparent base
class, using just a couple of class attributes from the sub-classes
to drive the work. NullResponse can also be handled within this scheme.

This gets rid of lots of trivial classes that weren't contributing terribly much.
And replaces quite a lot of code in the various Transparent classes with
just setting a few class attributes.

Framer now directly chooses one of (Heartbeat|Transparent)(Request|Response)
which gets rid of (Client|Server)IncomingMessage layer, but does leak a tiny bit
of encapsulation out of pdu. But I anticipate moving framer into pdu/ in which
case that won't matter.

Also gets rid of the 'expected_response' stuff, which is used by the
client to match a Response to the originating Request. Previously,
the code was actually creating an appropriate TransparentResponse
instance to match the response, and then waiting for an incoming
message of that shape. But if we simply define that the "shape_hash"
of the Request and Response are the same, we can just use the
original request rather than creating the anticipated response explictly.
(Heartbeat still has expected_response() since client uses that to create the reply.)

Internally, WriteHoldingRegister messages now use base_register
and register_values[], like the others, rather than register and value. Use some
rather ugly code to provide the familiar register + value interface.

TransparentMessage.__init__() now performs validation, so there's no need for
ensure_valid_state()

Simplify code to add the crc to the message - previously it was doing
the byte-swapping itself to make it little-endian, but since struct.pack
can do that directly, just add an add_16bit_le() method.

Note that when encoding a Response message, the CRC will probably be wrong.
We don't know the required algorithm, but fortunately we currently only ever
need to be able to encode requests.