Unrot the pytest support under tests/
Biggest regressions are
 - loss of pydantic
 - changing the crc calculation
 - framer no longer async

We've added a lot of new inverter registers, so the assertions
that the dictionary of register values fail because of all those
extra values. For now, I just check the subset defined by the test.
Also, some registers have been renamed.

Some minor changes to the library itself - not quite sure if
it's worth changing features just to fit the tests, but...
- allow parameters to Plant ctor
- add a RegisterGetter.getall() to enumerate all the registers
- add a RegisterGetter.__str__() to get a string representation (rather pointless)

Remove some unused conversion functions from register_cache.py
- I'm guessing they used to be used, but have now been replaced
  by the conversion functions in RegisterDef

ctor for WriteHoldingRegisterRequest() no longer throws an exception
based on whether it thinks the register is writable.