Provide an access method client.commands
First small step in a migration of the way commands are accessed.
Currently, they are just a bunch of static methods, unaware of
what sort of device the client is attached to. This means there's
no way to influence things like which commands are actually
available/safe for the remote device, and which is the preferred
slave address to access it.

The long term plan is that client.commands will return an
instance of a class which has access to the plant. For now,
this is a trivial first step where
  commands = client.commands
is equivalent to
  import givenergy_modbus.client.commands as commands

So applications can start migrating to the preferred syntax
immediately, and won't break when the implementation is changed.

Also, since some type declarations from specific 'list' to more
generic 'sequence', which means more efficient tuples can also be
pased. And run it through 'ruff format'.