Clone of Britkat Dev Build 21-07-2024
This is a clone of the Dev build Repo found inside Britkats DockerHub image as of 21-07-2024. 
Various changes to look into and look at merging as needed before the libraries splinter much further. 
Note: Only changes from raw code found in original image are in regards to Linefeed (all back to Unix) so changes are easier to follow.